clima-prob-backend/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
gunicorn==22.0.0
pydantic-settings==2.4.0
python-dotenv==1.0.1

# Ciencia de datos / IO
numpy==1.26.4
pandas==2.2.3
xarray==2024.7.0
netCDF4==1.7.1
cftime==1.6.4.post1
bottleneck==1.4.0
dask==2024.8.2
scipy==1.13.1
requests==2.32.3
pydap==3.4.1
earthaccess==0.15.1   # << actualizado

# ML
scikit-learn==1.5.2
statsmodels==0.14.2

# Cache / logs
redis==5.0.8
diskcache==5.6.3
loguru==0.7.2
